+++ sys/dev/iwx/if_iwx_debug.c
@@ -163,22 +163,24 @@ iwx_bbl_print_entry(struct iwx_bbl_entry *e)
 	case IWX_BBL_PKT_TX:
 		printf("pkt     ");
 		printf("seq %08d\t pkt len %ld",
-			e->seq, e->code);
+			e->seq, (long)e->code);
 		break;
 		printf("pkt dup ");
 		printf("seq %08d\t dup count %ld",
-			e->seq, e->code);
+			e->seq, (long)e->code);
 		break;
 	case IWX_BBL_CMD_TX:
 		printf("tx ->   ");
 		printf("seq %08d\tcode 0x%08lx (%s:%s)",
-			e->seq, e->code, get_label(command_group, group),
+			e->seq, (unsigned long)e->code,
+			get_label(command_group, group),
 			get_label(get_table(group), opcode));
 		break;
 	case IWX_BBL_CMD_RX:
 		printf("rx      ");
 		printf("seq %08d\tcode 0x%08lx (%s:%s)",
-			e->seq, e->code, get_label(command_group, group),
+			e->seq, (unsigned long)e->code,
+			get_label(command_group, group),
 			get_label(get_table(group), opcode));
 		break;
 	}
